+++ CHANGES.rst
@@ -56,6 +56,13 @@ Changelog
 - Added `utils.get_annexes_config` function to be able to get the annexes config
   depending on `context` and annex `portal_type`.
   [gbastien]
+- Adapted `ToolPloneMeeting._users_groups_value` returned value and cachekey:
+  - before we returned the full users/groups association which may be huge and
+    take much RAM, now we only return md5 hash;
+  - before the cachekey was for one request now we use the PAS principal
+    added/removed from from to invalidate cache.
+
+  [gbastien]
 
 4.2b18 (2021-10-13)
 -------------------